Introduce new naming convention for Equihash
Name (n, k) Params

Equihash16x5 (96, 5)
Equihash20x9 (200, 9) = Equihash
Equihash21x9 (210, 9)
Equihash24x5 (144,5)
Equihash24x7 (192,7)

The new numbers describe the number of bits matched in each round for the algorithm, and provide a relative measure of the “hardness” of the algorithm. (You can calculate the number from the n and k as so: n/(k+1) )

Because these numbers lend themselves to sorting in the order of memory “hardness,” it’s a natural way to show the progression - the progression is not readily apparent in the (n, k) parameters.