Update miner FireIce/XmrStak to v2.5.2
- RainbowMiner nodevfee version
This release fixes a NVIDIA bug which reduce the hash rate for monero/cryptonight_v8 a little bit.
The upcoming Graft fork on 31st Oct to the POW cryptonight_v8 is supported with this release.

Config files from 2.5.X are compatible and 2.4.X or older are NOT compatible to this release.

Changelog:
NVIDIA: fix wrong number of threads #2003
Update for upcoming Graft CNv2 fork #2001