Update miner TTminer to v2.1.16 (nvidia)
   - Reduced number of 'Duplicate shares' for ProgPoW algo
   - Fixed a bug that I added in version 2.1.15 that causes the hashrate to go down 50% for ProgPoW
   - Added intensity command line parameter
   - Added grid size command line parameter
   - Another Performance improvement for MTP

This version allow to set the intensity or the grid-size for each GPU in your rig. The miner shows the active intensity on mining start fo4r each GPU. You can overwrite this value using the -i (or -mi, -intensity) command line parameter. If you want to have a more detailed control over the grid-size you can use the intensity grid/grid-size command line parameter -ig (or -gs). The number you specify is the grid size. On mining start the miner shows you the active grid size. Take this value as a starting point for a more optimized setting for you system. See readme for more information - or ask me.