Update miner FireIce to v2.10.2 (amd,cpu,nvidia)
This release fixes a memory leak in the OpenCL (AMD) backend if cryptonight_r is mined which filled the full main memory and crashed the miner after a few hours.
Please update your miner if you are mining cryptonight_r, monero or cryptonight_v8.
The new option --amdCacheDir allows to select the folder for the OpenCL binaries.

Changelog:
- add cli option amdCacheDir #2332
- Repair OSX CUDA searched-path for NVRTC #2335
- OpenCL: fix memory leak #2345