Update miner MiniZ to v1.4o (nvidia)
Added support for 125,4 (ZelCash).
Added support for 210,9 (Aion). (Testing!)
Fixed solution/iteration for 96,5 (Testing!)
Minor bug fixes.
Keep in mind that 210,9 (Aion) is still not fully tested and optimized. We will keep working on it.

For the next release we will

work on the support for the new Beam algo/fork;
revise all the currently supported algos to see if we can add some improvements; Smiley
start looking at VDS support.