Update miner CcminerTcr to v1.2.6 (nvidia)
Procedure for pool mining:

ccminer -a mtp-tcr -o stratum+tcp://pool-mtp.tecracoin.io:4560 -u TPkxM1Aw872FL9gs4udCDzy5hAG7M7sVSE -p 0,d=0.032

This new version of ccminer also allows solo mining on mtp-tcr.

Procedure for solo mining.

wallet side
edit or create the file tecracoin.conf and add these informations to allow mining (this is an example):
server=1
rpcuser=djm34
rpcpassword=password
rpcport=8382
rpcallowip=127.0.0.1 (if the IP address of the miner is the same as the one of the wallet), otherwise it should be the address of the miner)

in the miner command line:
ccminer -a mtp-tcr -o http://127.0.0.1:8382 -u djm34 -p password --coinbase-addr TPkxM1Aw872FL9gs4udCDzy5hAG7M7sVSE --no-getwork --no-stratum -i 20

In the case of the wallet is on a different machine http:// should point to the IP of the machine where the wallet is located and in the wallet config, rpcallowip to the address of machine where the miner is located

The address given in --coinbase-addr, does not need to an address of the wallet to which the miner is connected any tecracoin address should work.

ATTENTION: for rpcuser and rpcpassword chose non obvious one especially if you are connecting outside of a private network as they could be used to steal your fund... (probably not with rpcallowip, although this one should be kept as restrictive as possible)