Update miner CcminerTcr to v1.2.8 (nvidia)
reduce cpu usage on latest nvidia drivers. The cpu should have at least 2 cores or 2 threads (if single core).
This version allows to have a stable 100% gpu usage on all gpu while using --cpu-affinity 1

You need to add to the ccminer command line: --cpu-affinity 1 (or 2 or 4).
This will assign all the mining threads to a single thread (it is mandatory that this thread is different from thread 0, hence the cpu-affinity)

The cpu usage will be the one you would get if your system had only one gpu.

This version may also improve the running on multigpu system (to confirm)