Update miner TTminer to v5.0.1 (nvidia)
I added a new error message that will help linux user to detect incomplete or missing cuda installation. Older versions said that the algo information is missing. That message is misleading on linux systems. The reason is in most cases that the nvidia builtins couldn't be loaded. They are required for TT.