Update miner Phoenix to v5.0e (amd,nvidia)
We added additional fixes and some workarounds for the ETH to ETC switching problem in the new version PhoenixMiner 5.0e, but with some Linux drivers, there is no full stability (e.g. 19.20 are crashing if the miner starts mining ETC then switches to ETH but not if it goes ETH, then ETC, then ETH, and so on). But even with 19.20, switching should work fine until about 5-7 DAG epochs before the RAM of 4 GB cards would not be enough for mining ETC.