Update miner Teamred to v0.7.6.3 (amd)
This is a special alpha release for Nimiq for early testing.

The main change is that the network proxy needed to run TRM is now integrated and executed automatically by the miner. It's also bundled in the release. This means you should now forget about the proxy and just run the miner with e.g. -o stratum+tcp://nimiq.icemining.ca:2053 and the miner will start/stop the proxy automatically.

Please read the packaged README and NIMIQ_MINING.txt and join the TRM discord for more info.