Update core
- add parameter "Region" to pools.config.txt:

All region stratums are chosen automatically, depending on your "Region" setting in config.txt. For some cases, you might want to select a specific region stratum for a pool.

Example:

    "Nicehash": {
        "BTC": "<YOUR_BTC_ADDRESS>",
        "Worker": "$WorkerName",
        "Region": "US"
    },

In this example, only the US stratum will be used for Nicehash. Make sure, that the pool supports a stratum for the selected region. If it doesn't, the pool will not show up anymore.