Update miner Trex to v0.24.4 (nvidia)
- (autolykos2) Increase pool-side hashrate by 1-4%.
The miner now generates the dataset for the next ERGO block before it arrives with a small penalty to reported hashrate, and when the next block does arrive, it immediately starts hashing without losing 1-4 seconds on creating the dataset.
Memory requirements are doubled in this mode as the GPU has to hold two memory buffers. If there is not enough memory on the GPU for two datasets, the miner will fall back to running in single buffer mode.
To take the full advantage of this mode it's recommended not to set power limit but instead limit power consumption with --lock-cclock - this will allow the miner to draw more power during dataset rebuild and not lose hashrate. If you prefer to keep your power consumption constant, you can still set the power limit, however, in that case you'll notice a slight hashrate drop after each block.
Please note that in both cases the miner should give you higher pool-side hashrate in the long run compared to previous versions. Miner's reported hashrate now fully accounts for dataset creation time.
- (kawpow, firopow) Report hashrate to the mining pool
- (WebUI) Add support for displaying multiple rigs (nodes) on the same page; various improvements and fixes
- Add --devices-info parameter to list available CUDA devices

Bug fixes:
- (ethash, autolykos2) LHR lock detection is triggered by short-term hashrate drops, e.g. when turning on the monitor connected to the GPU.
- Original overclock settings are not restored upon miner shutdown in dual mining mode (only applicable if you overclock GPUs with t-rex)
- (WebUI) Cannot set --gpu-report-interval to 0 (disable)
- (WebUI, API) GPUs are sometimes paused with a considerable delay (should be immediate)