Update miner Lolminer to v1.39 (amd,nvidia)
#### Feature Changes
- Added support for mining TON. Use --algo TON to mine it. Developer fee is 1.5%, AMD cards since GCN1 (via OpenCL) and Nvidia cards Maxwell and newer generation (via Cuda) are supported.
- Added support for mining Ubiq. Use --algo UBQHASH to mine it. The fee is 0.7% as with all Ethash based coins.
- Added support for routing your stratum traffic through a socks5 proxy server. Use --socks5 : to enable it. Note: at the moment this does not work with https traffic as in the TON algorithm. We will work on that.

#### Fixes
- Added a timeout for DOH requests, so they can no longer hang indefinitely.
- Fixed ethash support for RX 5500 series GPUs on more recent drivers.

#### Notes about TON pool connection
At the moment only the https pool communication protocol is supported. Therefore your --pool address must start with https://. This protocol is not ideal for mining, but we work on establishing a real stratum protocol with some pools in the near future, stay tuned!
Note when using a pool address starting with https:// specifying a port is optional (the default 443 is used when not specifying one). Also when mining TON via https lolMiner automatically uses TLS 1.2 encryption with skipped host name verification. As a final notice: always make sure your TON pool either does not need miner wallet registration or you registered - many popular pools require to register your wallet address to mine it.

#### Notes about TON performance and tuning
With regard to tuning for TON: The algorithm needs core, core and more core. Therefore maximizing the core clock within the limits of you cooling and energy supply is the key for the ideal performance. It might be beneficial to reduce the core voltage to produce less heat and less energy consumption - this way the cards keep more distance to their thermal limits and can often clock higher as well. We recommend this algorithm on your AMD RX 6000 cards!