Update miner Lolminer to v1.58 (amd,nvidia)
- Added Kaspa only mining mode for Nvidia Pascal and new and AMD Polaris, Navi and Big Navi. Use --algo KASPA to mine it. Fee is 0.75%. (issue #2158)

Tuning advice: On Nvidia Turing and Ampere GPUs this algorithm works best when locking the core clock to desired value (higher gives more performance, lower reduces power draw). We recommend to not set the power limit. Also to save energy, the memory clock can be chosen rather low. On Linux you can use --mclk 810 (but when you change back to an other algorithm you might need to reset this or reboot the rig).