Update miner LolMiner to v1.69 (amd,nvidia)
- Added Nexa support for AMD Vega / VII (1) and RDNA 1-3 GPUs (2)
- The parameter --keepfree can now be used to keep a certain amount of memory free on each GPU on Nexa mining. This might end up in allocating a smaller lookup table, so the GPU memory has space left for other workloads (3)
- The --keepfree parameter is now working for each GPU separately. Use a comma separated list of values if you want to assign different values for each GPU.
- Nexa mining will now allocate memory for each GPU sequentially and delay later cards startup slightly to help systems with small virtual memory.

Notes
(1) AMD Vega and VII are currently only supported on GPU drivers running the Rocm compute OpenCL driver stack. This is e.g. default for Vega GPUs on the latest Linux drivers, but not on pre 2021 drivers. Support for older drivers (and maybe older GPUs) might be included with the upcoming releases.
(2) Attention: we have seen cases where one or multiple RX 6000 GPUs have been mining significantly slower then identical cards on the same rig. This happens only on older 21.x or 20.40 Linux drivers. We recommend to update to 21.50 or 22.20 to fix the issue.
(3) For users of the alternative Windows version of 1.68: Use --keepfree 1024 or higher to get the effect the alternative version had.

General note: The Nexa gpu codes are (and will be in foreseeable future) joint work with Iedoc from BzMiner. Miner fees will be evenly shared regardless of which of these two flavors of the code you prefer.