Update miner SrbMinerMulti to v2.3.6 (amd,cpu,intel,nvidia)
- Added support for algorithm 'dynex' on NVIDIA GPUs
- Performance improvement on algorithm 'dynex' on AMD GPUs
- Removed Dynex optimised dual kernels
- Lowered devfee for algorithm 'ethashb3' to standard 0.85%
- Added parameter '--busy-wait-recheck' which sets the NVIDIA/INTEL GPUs OpenCL busy wait recheck value per algorithm*

#### --busy-wait-recheck:
Nvidia's OpenCL implementation has a bug known for many years which manifests in the way of having higher CPU usage (LA - load average) than it should be.
This will probably never get fixed, and its present only in their OpenCL implementation, but not in CUDA.
SRBMiner uses OpenCL for NVIDIA, and a workaround for this bug is used since the addition of NVIDIA support, but doing so (lowering/fixing the CPU usage bug) comes with a penalty, which is lower hashrate in some cases.
Every rig behaves differently depending on the CPU and number of GPU's used on it.
Unfortunately lowering the CPU usage also might lower the hashrate too, so finding an optimal 'golden middle' is important.

--busy-wait-recheck accepts values from 0.01 to 0.99 -> lower the value, higher the CPU usage will be, which allows for higher hashrate

Default value of --busy-wait-recheck is 0.3, which might or might not be the best value for your rig.

As a first step, set '--busy-wait-recheck 0.01' to get the maximum hashrate, then increase the value of --busy-wait-recheck and do some tests.
Experiment and try out different values to find an acceptable hashrate / CPU LA ratio.