Update miner OneZeroMiner to v1.2.7 (nvidia)
This version is mainly to support dual mining Dynex and Zil.
Add support for Zilliqa(0% fee)
Add new options:
    --a2
    --o2
    --w2
    --p2
    --worker, --worker2
    --cclk2
    --coff2
    --mclk2
    --moff2
    --fan2
    --pl2
    --zil-before-script
    --zil-after-script
    --zil-test
    --zil-test-port
Add support for nvtool on hiveos for locking clocks on older gpus.
    The miner uses nvtool on hiveos as a fallback when the built-in oc functionality fails.

Note:
Please refer to the readme section of this repo or use --help on the miner for more documentation. There are also bat/sh files for mining DNX+ZIL in the packages.

Known issues:
If a zilliqa session is active when the miner is started, the first few dynex shares after switching back to it might get rejected. This is an edge case and should only happen once in the worst case scenario.